requirements.txt
aio-pika==8.2.4
aiormq==6.4.2
amqp==5.1.1
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.1
billiard==3.6.4.0
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==38.0.1
distro==1.8.0
Django==4.1.2
django-bootstrap-form==3.4
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==6.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.4
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==21.3
pamqp==3.2.1
paramiko==2.11.0
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2==2.9.4
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.6
PyYAML==5.4.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
texttable==1.6.4
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
yarl==1.8.1
zipp==3.10.0
uwsgi==2.0.21
model-bakery==1.9.0
